requirements.txt
aiofiles==0.7.0
aiosqlite==0.17.0
alembic==1.6.5
asgiref==3.4.1
asyncpg==0.23.0
click==8.0.1
ecdsa==0.17.0
fastapi==0.65.2
greenlet==1.1.0
h11==0.12.0
Jinja2==3.0.1
Mako==1.1.4
MarkupSafe==2.0.1
passlib==1.7.4
pyasn1==0.4.8
pydantic==1.8.2
python-dateutil==2.8.1
python-editor==1.0.4
python-jose==3.3.0
python-multipart==0.0.5
redis==3.5.3
rsa==4.7.2
six==1.16.0
SQLAlchemy==1.4.20
starlette==0.14.2
typing-extensions==3.10.0.0
uvicorn==0.14.0
websockets==9.1
